requirements.txt
cycler==0.10.0
future==0.18.2
imutils==0.5.3
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
opencv-python==4.1.2.30
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
torch==1.5.0
torchvision==0.6.0
